requirements/ml.txt
lxml==6.0.2
pandas==3.0.1
PyMuPDF==1.27.2
python-docx==1.2.0
requests==2.32.5
